chore: apply review - 2
